+++ app/src/main/java/com/yichiuan/weatherapp/data/openweathermap/model/Wind.java
@@ -8,7 +8,7 @@ import com.google.gson.TypeAdapter;
 @AutoValue
 public abstract class Wind {
 
-    public abstract int deg();
+    public abstract float deg();
 
     public abstract float speed();
